feat: port client app from next.js to pure react
