+++ client/.prettierrc
@@ -1,0 +1,27 @@
+{
+  "overrides": [
+    {
+      "files": ["*.js", "*.ts"],
+      "options": {
+        "printWidth": 100,
+        "semi": true,
+        "tabWidth": 2,
+        "useTabs": false,
+        "singleQuote": true,
+        "bracketSpacing": true
+      }
+    },
+    {
+      "files": ["*.js", "*.ts", "*.json"],
+      "options": {
+        "useTabs": false
+      }
+    },
+    {
+      "files": "*.md",
+      "options": {
+        "useTabs": false
+      }
+    }
+  ]
+}
