--- client/lib/assets.json
@@ -1,15 +1,0 @@
-[
-    {
-        "name": "ETH",
-        "color": "CoinETH",
-        "amount": 0,
-        "img": "/assets/eth.svg"
-    },
-    {
-        "name": "DAI",
-        "color": "CoinDAI",
-        "amount": 0,
-        "img": "/assets/dai.svg"
-    }
-    
-]
