--- client/pages/assets.tsx
@@ -1,7 +1,0 @@
-import Assets from "../components/Assets";
-
-export default function AssetsPage() {
-  return (
-    <Assets />
-  )
-}
