--- client/pages/setup.tsx
@@ -1,7 +1,0 @@
-import { Setup } from '../components/Setup'
-
-export default function SetupPage() {
-  return (
-    <Setup />
-  )
-}
