--- client/pages/swap.tsx
@@ -1,7 +1,0 @@
-import Swap from '../components/Swap'
-
-export default function SwapPage() {
-  return (
-    <Swap />
-  )
-}
