--- client/prettier.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-  arrowParens: 'always',
-  singleQuote: true,
-  tabWidth: 2,
-  semi: false,
-}
