+++ client/public/fuel-logo-192x192.png
