+++ client/public/fuel-logo-512x512.png
